Separate HIP sketching kernels from handler kernels.
