+++ resolve/hip/CMakeLists.txt
@@ -8,12 +8,14 @@
 
 set(ReSolve_HIP_SRC
     hipKernels.hip
+    hipSketchingKernels.hip
     hipVectorKernels.hip
     MemoryUtils.hip
 )
 
 set(ReSolve_HIP_HEADER_INSTALL
     hipKernels.h
+    hipSketchingKernels.h
     hipVectorKernels.h
     HipMemory.hpp
     hip_check_errors.hpp
