+++ resolve/random/RandSketchingFWHT.cpp
@@ -6,10 +6,10 @@
 #include <resolve/vector/Vector.hpp>
 #include <resolve/utilities/logger/Logger.hpp>
 #ifndef RESOLVE_USE_GPU
-#include <resolve/cpu/cpuKernels.h>
+#include <resolve/cpu/cpuSketchingKernels.h>
 #endif
 #ifdef RESOLVE_USE_HIP
-#include <resolve/hip/hipKernels.h>
+#include <resolve/hip/hipSketchingKernels.h>
 #endif
 #ifdef RESOLVE_USE_CUDA
 #include <resolve/cuda/cudaKernels.h>
